Add documentation comments
