+++ app/lib/api/client.ts
@@ -1,7 +1,7 @@
 import { isTokenExpired } from "../auth/oauth-client";
 import { refreshAccessToken } from "../auth/oauth-server";
 import { setToken } from "../auth/tokens";
-import { substituteObject, substituteVariables, Token } from "../workflow";
+import { substituteObject, substituteVariables, Token, Endpoint, LogEntry } from "../workflow";
 import {
   CONNECTION_IDENTIFIERS,
   HTTP_METHODS_WITH_BODY,
@@ -10,6 +10,91 @@ import {
   WORKFLOW_CONSTANTS,
 } from "../workflow/constants";
 import { ApiRequestOptions } from "./types";
+
+function getTokenForConnection(
+  endpoint: Endpoint,
+  tokens: { google?: Token; microsoft?: Token },
+  _onLog?: (entry: LogEntry) => void,
+): { token: Token | null; provider: Provider | null } {
+  let token: Token | null = null;
+  let provider: Provider | null = null;
+  if (
+    endpoint.conn.includes(CONNECTION_IDENTIFIERS.GOOGLE) ||
+    endpoint.conn.includes(CONNECTION_IDENTIFIERS.GOOGLE_CI)
+  ) {
+    token = tokens.google ?? null;
+    provider = PROVIDERS.GOOGLE;
+  } else if (endpoint.conn.includes(CONNECTION_IDENTIFIERS.MICROSOFT)) {
+    token = tokens.microsoft ?? null;
+    provider = PROVIDERS.MICROSOFT;
+  }
+  return { token, provider };
+}
+
+async function refreshTokenIfNeeded(
+  token: Token,
+  provider: Provider,
+  onLog?: (entry: LogEntry) => void,
+): Promise<Token> {
+  if (!isTokenExpired(token) || !token.refreshToken) {
+    return token;
+  }
+
+  let refreshAttempts = 0;
+  while (refreshAttempts < WORKFLOW_CONSTANTS.MAX_REFRESH_ATTEMPTS) {
+    try {
+      onLog?.({
+        timestamp: Date.now(),
+        level: "info",
+        message: `Refreshing token for ${provider}, attempt ${refreshAttempts + 1}`,
+      });
+      const refreshedToken = await refreshAccessToken(
+        provider,
+        token.refreshToken,
+      );
+      await setToken(provider, refreshedToken);
+      return refreshedToken;
+    } catch (error) {
+      refreshAttempts++;
+      onLog?.({
+        timestamp: Date.now(),
+        level: "error",
+        message: `Token refresh attempt ${refreshAttempts} failed for ${provider}`,
+        data: error,
+      });
+      if (refreshAttempts === WORKFLOW_CONSTANTS.MAX_REFRESH_ATTEMPTS) {
+        throw new Error(
+          `${PROVIDERS[provider.toUpperCase() as keyof typeof PROVIDERS]} authentication expired. Please re-authenticate.`,
+        );
+      }
+    }
+  }
+  return token;
+}
+
+function buildAuthenticatedRequest(
+  connection: { auth: string },
+  endpoint: Endpoint,
+  token: Token,
+  body: unknown,
+): RequestInit {
+  const headers: Record<string, string> = {
+    "Content-Type": "application/json",
+  };
+  const authHeader = connection.auth
+    .replace("{googleAccessToken}", token.accessToken)
+    .replace("{azureAccessToken}", token.accessToken);
+  headers["Authorization"] = authHeader;
+
+  const requestOptions: RequestInit = {
+    method: endpoint.method,
+    headers,
+  };
+  if (body && HTTP_METHODS_WITH_BODY.includes(endpoint.method as (typeof HTTP_METHODS_WITH_BODY)[number])) {
+    requestOptions.body = JSON.stringify(body);
+  }
+  return requestOptions;
+}
 
 async function handlePublicRequest(
   options: ApiRequestOptions,
@@ -70,58 +155,17 @@ async function handleAuthenticatedRequest(
   const connection = connections[endpoint.conn];
   const capturedValues: Record<string, string> = {};
 
-  let token: Token | null = null;
-  let provider: Provider | null = null;
+  const { token: initialToken, provider } = getTokenForConnection(
+    endpoint,
+    tokens,
+    onLog,
+  );
 
-  if (
-    endpoint.conn.includes(CONNECTION_IDENTIFIERS.GOOGLE) ||
-    endpoint.conn.includes(CONNECTION_IDENTIFIERS.GOOGLE_CI)
-  ) {
-    token = tokens.google ?? null;
-    provider = PROVIDERS.GOOGLE;
-  } else if (endpoint.conn.includes(CONNECTION_IDENTIFIERS.MICROSOFT)) {
-    token = tokens.microsoft ?? null;
-    provider = PROVIDERS.MICROSOFT;
-  }
-
-  if (token == null) {
+  if (!initialToken || !provider) {
     throw new Error(`No token available for connection: ${endpoint.conn}`);
   }
 
-  // Check and refresh token if needed
-  if (isTokenExpired(token) && token.refreshToken && provider) {
-    let refreshAttempts = 0;
-
-    while (refreshAttempts < WORKFLOW_CONSTANTS.MAX_REFRESH_ATTEMPTS) {
-      try {
-        onLog?.({
-          timestamp: Date.now(),
-          level: "info",
-          message: `Refreshing token for ${provider}, attempt ${refreshAttempts + 1}`,
-        });
-        const refreshedToken = await refreshAccessToken(
-          provider,
-          token.refreshToken!,
-        );
-        await setToken(provider, refreshedToken);
-        token = refreshedToken;
-        break;
-      } catch (error) {
-        refreshAttempts++;
-        onLog?.({
-          timestamp: Date.now(),
-          level: "error",
-          message: `Token refresh attempt ${refreshAttempts} failed for ${provider}`,
-          data: error,
-        });
-        if (refreshAttempts === WORKFLOW_CONSTANTS.MAX_REFRESH_ATTEMPTS) {
-          throw new Error(
-            `${PROVIDERS[provider.toUpperCase() as keyof typeof PROVIDERS]} authentication expired. Please re-authenticate.`,
-          );
-        }
-      }
-    }
-  }
+  const token = await refreshTokenIfNeeded(initialToken, provider, onLog);
 
   const path = substituteVariables(endpoint.path, variables, {
     throwOnMissing: throwOnMissingVars,
@@ -136,20 +180,6 @@ async function handleAuthenticatedRequest(
     url += `?${params.toString()}`;
   }
 
-  const headers: Record<string, string> = {
-    "Content-Type": "application/json",
-  };
-
-  const authHeader = connection.auth
-    .replace("{googleAccessToken}", token.accessToken)
-    .replace("{azureAccessToken}", token.accessToken);
-  headers["Authorization"] = authHeader;
-
-  const requestOptions: RequestInit = {
-    method: endpoint.method,
-    headers,
-  };
-
   const finalBody = body
     ? substituteObject(body, variables, {
         throwOnMissing: throwOnMissingVars,
@@ -157,14 +187,12 @@ async function handleAuthenticatedRequest(
       })
     : undefined;
 
-  if (
-    finalBody &&
-    HTTP_METHODS_WITH_BODY.includes(
-      endpoint.method as (typeof HTTP_METHODS_WITH_BODY)[number],
-    )
-  ) {
-    requestOptions.body = JSON.stringify(finalBody);
-  }
+  const requestOptions = buildAuthenticatedRequest(
+    connection,
+    endpoint,
+    token,
+    finalBody,
+  );
 
   const response = await fetch(url, requestOptions);
